+++ locale/nl/index.js
@@ -8,6 +8,7 @@ module.exports = {
 	narrow: locale.narrow,
 	// Additional styles.
 	'long-time': require('../../locale-more-styles/nl/long-time.json'),
+	'now': require('../../locale-more-styles/nl/now.json'),
 	// Quantifier.
 	quantify: locale.quantify
 }